Rescue redirect responses with no “Location” header
Browsers, in accordance with RFC 7231, treat redirects with no “Location” header as 200 responses, ie. show the response body, which in theory would contain a clickable link to the redirected resource… I don’t think this is a sensible behavior for an HTTP library and probably just means we are dealing with a poorly configured web server anyway.